aus dem menu controller mehrere kleinere controller gemacht
aus dem menu controller mehrere kleinere controller gemacht, außer user,
w-eingang und ausgang hat jedes model nun einen eigenen controller
